+++ scripts/pre_extra_script.py
@@ -19,3 +19,4 @@ def process_node(node):
     return node
 
 env.AddBuildMiddleware(process_node)
+