+++ src/dimmer_version.h
@@ -1,0 +1,4 @@
+// library.json { "version":"2.2.3" }
+#define DIMMER_VERSION_MAJOR 2
+#define DIMMER_VERSION_MINOR 2
+#define DIMMER_VERSION_REVISION 3